Move docker_async_source to its own source file
