+++ userspace/libsinsp/parsers.cpp
@@ -4741,7 +4741,7 @@ void sinsp_parser::parse_container_json_evt(sinsp_evt *evt)
 		}
 
 #if !defined(MINIMAL_BUILD) && !defined(_WIN32)
-		libsinsp::container_engine::docker::parse_json_mounts(container["Mounts"], container_info->m_mounts);
+		libsinsp::container_engine::docker_async_source::parse_json_mounts(container["Mounts"], container_info->m_mounts);
 #endif
 
 		sinsp_container_info::container_health_probe::parse_health_probes(container, container_info->m_health_probes);
